# backend/docs/swagger.yaml
basePath: /api/v1
...
info:
  contact: {}
  description: Backend API for the Mihombreng controller
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Mihombreng API
  version: "1.0"
paths:
...
  /mihomo/{dir}:
    get:
      description: Get list of all files in the specified directory
      parameters:
      - description: Directory name (configs, proxy_providers, rule_providers)
        in: path
        name: dir
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of files
          schema:
            items:
              type: string
            type: array
        "400":
          description: Error message
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error message
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get files from directory
      tags:
      - Mihomo files
    post:
      consumes:
      - application/json
      description: Create a new file with the specified content
      parameters:
      - description: Directory name (configs, proxy_providers, rule_providers)
        in: path
        name: dir
        required: true
        type: string
      - description: Filename and content
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Error message
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error message
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new file
      tags:
      - Mihomo files
  /mihomo/{dir}/{filename}:
    delete:
      description: Delete an existing file
      parameters:
      - description: Directory name (configs, proxy_providers, rule_providers)
        in: path
        name: dir
        required: true
        type: string
      - description: Filename of the file to delete
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Error message
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error message
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a file
      tags:
      - Mihomo files
    get:
      description: Get the content of a specific file
      parameters:
      - description: Directory name (configs, proxy_providers, rule_providers)
        in: path
        name: dir
        required: true
        type: string
      - description: Filename of the file
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: File content
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Error message
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error message
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get file content
      tags:
      - Mihomo files
    put:
      consumes:
      - application/json
      description: Update the content of an existing file
      parameters:
      - description: Directory name (configs, proxy_providers, rule_providers)
        in: path
        name: dir
        required: true
        type: string
      - description: Filename of the file
        in: path
        name: filename
        required: true
        type: string
      - description: New content
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Error message
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error message
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a file
      tags:
      - Mihomo files
  /mihomo/{dir}/{filename}/download:
    get:
      description: Download a specific file
      parameters:
      - description: Directory name (configs, proxy_providers, rule_providers)
        in: path
        name: dir
        required: true
        type: string
      - description: Filename of the file
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File content
          schema:
            type: file
        "400":
          description: Error message
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error message
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Download a file
      tags:
      - Mihomo files
  /mihomo/{dir}/{filename}/rename:
    put:
      consumes:
      - application/json
      description: Rename an existing file
      parameters:
      - description: Directory name (configs, proxy_providers, rule_providers)
        in: path
        name: dir
        required: true
        type: string
      - description: Current filename of the file
        in: path
        name: filename
        required: true
        type: string
      - description: New filename
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Error message
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error message
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Rename a file
      tags:
      - Mihomo files
  /mihomo/{dir}/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload a new file
      parameters:
      - description: Directory name (configs, proxy_providers, rule_providers)
        in: path
        name: dir
        required: true
        type: string
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Error message
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error message
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Upload a file
      tags:
      - Mihomo files
  /mihomo/active-config:
    get:
      description: Get the currently active mihomo configuration file path
      produces:
      - application/json
      responses:
        "200":
          description: Active config path
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error message
          schema:
            additionalProperties: true
            type: object
      summary: Get active config path
      tags:
      - Mihomo files
    put:
      consumes:
      - application/json
      description: Set the active mihomo configuration file path
      parameters:
      - description: Filename to set as active
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Error message
          schema:
            additionalProperties: true
            type: object
        "404":
          description: File not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error message
          schema:
            additionalProperties: true
            type: object
      summary: Set active config path
      tags:
      - Mihomo files
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"